--- apk/org.wso2.apk.apimgt.rest.api.backoffice.v1.common/src/main/java/org/wso2/apk/apimgt/rest/api/backoffice/RestApplication.java
@@ -1,9 +1,0 @@
-package org.wso2.apk.apimgt.rest.api.backoffice;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("/SanojPunchihewa/BackOfficeAPI/1.0.0")
-public class RestApplication extends Application {
-  // Add implementation-specific details here
-}
