+++ apk/org.wso2.apk.apimgt.rest.api.backoffice.v1.common/src/main/java/org/wso2/apk/apimgt/rest/api/backoffice/v1/common/impl/ApisApiCommonImpl.java
@@ -58,14 +58,14 @@ import org.json.JSONTokener;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.CertificateRestApiUtils;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.CommentMappingUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.DocumentationMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.CertificateRestApiUtils;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.CommentMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.DocumentationMappingUtil;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.crypto.CryptoTool;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.crypto.CryptoToolException;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.crypto.CryptoToolUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.APIMappingUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.PublisherCommonUtils;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.APIMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.PublisherCommonUtils;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.APIDTO;
 import org.wso2.apk.apimgt.api.APIDefinition;
 import org.wso2.apk.apimgt.api.APIDefinitionValidationResponse;
@@ -131,10 +131,10 @@ import org.wso2.apk.apimgt.impl.wsdl.util.SequenceUtils;
 import org.wso2.apk.apimgt.rest.api.common.RestApiCommonUtil;
 import org.wso2.apk.apimgt.rest.api.common.RestApiConstants;
 import org.wso2.apk.apimgt.rest.api.common.dto.ErrorDTO;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.CertificateMappingUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.ExternalStoreMappingUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.GraphqlQueryAnalysisMappingUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.OperationPolicyMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.CertificateMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.ExternalStoreMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.GraphqlQueryAnalysisMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.OperationPolicyMappingUtil;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.APIExternalStoreListDTO;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.APIKeyDTO;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.APIMonetizationInfoDTO;
