+++ apk/org.wso2.apk.apimgt.rest.api.backoffice.v1.common/src/main/java/org/wso2/apk/apimgt/rest/api/backoffice/v1/common/impl/EndpointCertificatesApiCommonImpl.java
@@ -29,11 +29,11 @@ import org.wso2.apk.apimgt.api.dto.CertificateMetadataDTO;
 import org.wso2.apk.apimgt.api.model.APISearchResult;
 import org.wso2.apk.apimgt.impl.certificatemgt.ResponseCode;
 import org.wso2.apk.apimgt.impl.utils.APIUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.CertificateRestApiUtils;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.CertificateRestApiUtils;
 import org.wso2.apk.apimgt.rest.api.common.RestApiCommonUtil;
 import org.wso2.apk.apimgt.rest.api.common.RestApiConstants;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.APIMappingUtil;
-import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.CertificateMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.APIMappingUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.CertificateMappingUtil;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.APIMetadataListDTO;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.CertMetadataDTO;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.CertificateInfoDTO;
