+++ apk/org.wso2.apk.apimgt.rest.api.backoffice.v1.common/src/main/java/org/wso2/apk/apimgt/rest/api/backoffice/v1/common/impl/RuntimeApisApiCommonImpl.java
@@ -72,6 +72,7 @@ import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings.*;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.crypto.CryptoTool;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.crypto.CryptoToolException;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.crypto.CryptoToolUtil;
+import org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings.*;
 import org.wso2.apk.apimgt.rest.api.backoffice.v1.dto.*;
 import org.wso2.apk.apimgt.rest.api.common.RestApiCommonUtil;
 import org.wso2.apk.apimgt.rest.api.common.RestApiConstants;
