+++ apk/org.wso2.apk.apimgt.rest.api.backoffice.v1.common/src/main/java/org/wso2/apk/apimgt/rest/api/backoffice/v1/common/utils/mappings/APIMappingUtil.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings;
+package org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
