+++ apk/org.wso2.apk.apimgt.rest.api.backoffice.v1.common/src/main/java/org/wso2/apk/apimgt/rest/api/backoffice/v1/common/utils/mappings/EnvironmentMappingUtil.java
@@ -18,7 +18,7 @@
  * /
  */
 
-package org.wso2.apk.apimgt.rest.api.backoffice.v1.common.mappings;
+package org.wso2.apk.apimgt.rest.api.backoffice.v1.common.utils.mappings;
 
 import org.wso2.apk.apimgt.api.model.AsyncProtocolEndpoint;
 import org.wso2.apk.apimgt.api.model.Environment;
